feat(gameStates and gameFunctions modular functions)
gameStates will contain the two variables that determine if the game started or ended. Whereas gameFunction contains all the possible winning combinations and the function of checkWinner that contrast those with the ones chosed by both the player and the opponent, declaring a winner if they match or a tie if they dont.